+++ packages/terra-alert/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+4.32.0 - (June 2, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.31.0 - (May 26, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.30.0 - (May 19, 2020)
 ------------------
 ### Changed