+++ packages/terra-button-group/CHANGELOG.md
@@ -4,6 +4,12 @@ ChangeLog
 Unreleased
 ----------
 
+3.40.0 - (June 2, 2020)
+------------------
+### Fixed
+* Fixed `ButtonGroup.Button` not losing `focus` styles when made disabled via user interaction
+* Fixed `ButtonGroup.Button` not losing `is-selected` styles when made disabled via user interaction
+
 3.39.0 - (May 19, 2020)
 ------------------
 ### Changed