+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+3.39.0 - (June 2, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.38.0 - (May 26, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.37.0 - (May 19, 2020)
 ------------------
 ### Changed