+++ packages/terra-grid/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+6.21.0 - (June 2, 2020)
+------------------
+### Changed
+* Apply themes by context and directly include packaged themes.
+
+6.20.0 - (May 26, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 6.19.0 - (April 28, 2020)
 ------------------
 ### Changed