+++ packages/terra-overlay/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.49.0 - (June 2, 2020)
+------------------
+### Changed
+* Apply themes by context and directly include packaged themes.
+
 3.48.0 - (May 19, 2020)
 ------------------
 ### Changed