+++ packages/terra-profile-image/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.30.0 - (June 2, 2020)
+------------------
+### Changed
+* Apply themes by context and directly include packaged themes.
+
 3.29.0 - (April 28, 2020)
 ------------------
 ### Changed