+++ packages/terra-signature/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Apply themes by context and directly include packaged themes.
 
 2.29.0 - (April 28, 2020)
 ------------------