+++ packages/terra-spacer/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Apply themes by context and directly include packaged themes.
+
+3.39.0 - (June 2, 2020)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.38.0 - (May 19, 2020)
 ------------------